Add support for local paths as --url options (#3)
As Git can clone a local path in addition to a remote URL add support for using local paths for the --url option. If a local path is used for the --url option then there is no guarantee that the default branch created will be named `master`. To guard against this use `origin/master` instead (which is reasonable as `origin` is a very standard name for the default remote).

Additionally fix a bug where Git repo values were being accessed erroneously before the repo was cloned.